requirements.txt
# Core dependencies
pygame>=2.5.0
PyYAML>=6.0
requests>=2.31.0
Pillow>=10.0.0  # For sprite sheet creation script

# GPIO (for Raspberry Pi)
RPi.GPIO>=0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"

# Optional: For enhanced network scanning
python-nmap>=0.7.1

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

